Removed auto-dispose of ResizableController
