+++ lib/src/utils.dart
@@ -1,0 +1,8 @@
+/// Returns the sum of a list.
+num sum(List<num> list) {
+  num result = 0;
+  for (final element in list) {
+    result += element;
+  }
+  return result;
+}
